+++ packages/flame_jenny/jenny/lib/jenny.dart
@@ -1,4 +1,5 @@
-export 'src/character_storage.dart' show Character, CharacterStorage;
+export 'src/character.dart' show Character;
+export 'src/character_storage.dart' show CharacterStorage;
 export 'src/dialogue_runner.dart' show DialogueRunner;
 export 'src/dialogue_view.dart' show DialogueView;
 export 'src/errors.dart' show SyntaxError, NameError, TypeError, DialogueError;
